github-identicons/requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.8
matplotlib==3.10.5
numpy==2.3.4
packaging==25.0
pandas==2.3.1
pillow==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tkinter==3.13.9
tzdata==2025.2
